refactor: migrate to status_t enum, modularize lexer, and clean up parser/gen
**What changed:**

- **Error handling:** Replaced magic `int` return codes (`ERROR = 84`,
  `EINP`, `EPAR`, etc.) with a proper `status_t` enum (`SUCCESS`,
  `ELEX`, `EPARSE`, `EGEN`, `EMEM`, `EIMP`). Added `print_error()`
  returning `void *` for cleaner null-propagation in allocation
  failures.

- **Lexer refactor:** `lexer_create` now takes the input string
  directly. `lexer_run` no longer takes a buffer argument — the lexer
  owns its input. Removed monolithic `lexer_tokenize.c` and
  `get_tok_type.c` in favor of modular extraction. Added
  `lexer_skip_ignored`, `lexer_push_token`, and optional
  `MNCC_LEXER_DEBUG` output.

- **Parser cleanup:** Renamed `parse_sized_expression` →
  `parse_expression_with_size`, `parse_operator` →
  `parse_expression_with_operator`, `tokens_to_operator` →
  `token_to_operator`. Removed `match_operator.c` and
  `parse_operator.c`, replacing multi-token operator matching with
  single-token mapping. Added `get_expression_size`,
  `expression_pre_clean`, `get_operator_offset`. Optional
  `MNCC_PARSER_DEBUG` output. `TOK_EQ` is now `==`, `TOK_DEF` is `=`.

- **Gen:** Extracted `gen_get_var_offset` as a shared helper, removing
  duplicated `get_offset` statics across `gen_value.c`,
  `gen_declaration.c`, `gen_assignement.c`. Deleted
  `gen_get_registers.c` and `gen_header.c` (presumably inlined
  elsewhere).

- **Array:** `int` indexes and counts replaced with `size_t`/`ssize_t`.
  `array_destroy` takes `bool`. Removed `format_array_index`.
  `ARRAY_INITIAL_CAPACITY` constant added.

- **Misc:** Removed `test.c`, `exec.c` (dead code). Makefile gains
  `PARSER_DEBUG` and `LEXER_DEBUG` flags.

+++ src/parser/parser_moves.c
@@ -4,9 +4,10 @@
 ** File description:
 ** Parser token navigation and matching functions
 */
+#include <stddef.h>
 #include "parser/parser.h"
 
-token_t *parser_at(parser_t *p, int index)
+token_t *parser_at(parser_t *p, size_t index)
 {
     if (p->lexer->tokens->count <= index)
         return NULL;
@@ -36,7 +37,7 @@ bool parser_match(parser_t *p, token_type_t type)
 
 token_t **parser_tokens_peek(parser_t *p)
 {
-    int index = p->cursor;
+    size_t index = p->cursor;
 
     if (p->lexer->tokens->count <= index)
         return NULL;
